Fix loopcount_1_msec calculation on ColdFire
The 16-bit multiplication overflowed,
resulting to a huge value on the M5484LITE,
causing slowness specially when booting FreeMiNT.

On the FireBee, the truncated value was smaller,
but did't cause any visible trouble.

+++ bios/delay.c
@@ -121,7 +121,7 @@ void calibrate_delay(void)
         loopcount_1_msec = (loopcount * 24) / (intcount * 25);
 #else
   #if defined(MACHINE_FIREBEE) || defined(MACHINE_M548X)
-    loopcount_1_msec = cookie_mcf.sysbus_frequency * 1000;
+    loopcount_1_msec = (ULONG)cookie_mcf.sysbus_frequency * 1000;
   #endif
 #endif
 }
